initial state of the project
