--- src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Discount.java
@@ -1,30 +1,0 @@
-package pl.com.bottega.ecommerce.sales.domain.offer;
-
-import java.math.BigDecimal;
-
-public class Discount {
-    // discount
-    private String cause;
-    private Money money;
-
-    Discount(String discountCause){
-        this.cause = discountCause;
-        this.money = new Money();
-    }
-
-    public String getCause() {
-        return cause;
-    }
-
-    public void setCause(String cause) {
-        this.cause = cause;
-    }
-
-    public Money getMoney() {
-        return money;
-    }
-
-    public void setMoney(Money money) {
-        this.money = money;
-    }
-}
