--- src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Money.java
@@ -1,31 +1,0 @@
-package pl.com.bottega.ecommerce.sales.domain.offer;
-
-import java.math.BigDecimal;
-
-public class Money {
-    private String currency;
-    private BigDecimal value;
-
-    public Money(String currency,BigDecimal value)
-    {
-        this.value = value;
-        this.currency = currency;
-    }
-
-    public String getCurrency(){
-
-        return currency;
-    }
-
-    public void setCurrency(String currency) {
-        this.currency = currency;
-    }
-
-    public BigDecimal getValue() {
-        return value;
-    }
-
-    public void setValue(BigDecimal value) {
-        this.value = value;
-    }
-}
