--- src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Product.java
@@ -1,46 +1,0 @@
-package pl.com.bottega.ecommerce.sales.domain.offer;
-
-import java.math.BigDecimal;
-import java.util.Date;
-
-public class Product {
-
-    private String productId;
-    private String productName;
-    private Date productSnapshotDate;
-    private String productType;
-    private Money money;
-
-    public Product(String productId, String productName, Date productSnapshotDate, String productType)
-    {
-        this.productId = productId;
-        this.productName = productName;
-        this.productSnapshotDate = productSnapshotDate;
-        this.productType = productType;
-
-    }
-
-    public String getProductId() {
-        return productId;
-    }
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public Date getProductSnapshotDate() {
-        return productSnapshotDate;
-    }
-
-    public String getProductType() {
-        return productType;
-    }
-
-    public Money getMoney() {
-        return money;
-    }
-
-    public void setMoney(Money money) {
-        this.money = money;
-    }
-}
